# Conflicts:
#	api-repository/src/main/java/com/walking/api/repository/dao/member/PathFavoritesRepository.java
#	api-repository/src/main/java/com/walking/api/repository/dao/member/TrafficFavoritesRepository.java
#	api-repository/src/main/java/com/walking/api/repository/dao/traffic/TrafficDetailRepository.java
#	api-repository/src/main/java/com/walking/api/repository/dao/traffic/TrafficFavoritesRepository.java

+++ api/src/main/java/com/walking/api/web/handler/OptionalViewPointParamArgumentResolver.java
@@ -37,10 +37,10 @@ public class OptionalViewPointParamArgumentResolver implements HandlerMethodArgu
 		return OptionalViewPointParam.builder()
 				.viewPointParam(
 						ViewPointParam.builder()
-								.vblLat(Double.parseDouble(vblLat))
-								.vblLng(Double.parseDouble(vblLng))
-								.vtrLat(Double.parseDouble(vtrLat))
-								.vtrLng(Double.parseDouble(vtrLng))
+								.vblLat(Float.parseFloat(vblLat))
+								.vblLng(Float.parseFloat(vblLng))
+								.vtrLat(Float.parseFloat(vtrLat))
+								.vtrLng(Float.parseFloat(vtrLng))
 								.build())
 				.build();
 	}
